feature/refactor-easy-bugsnag
- Refactor EasyBugsnag
